SpindleStreamlitCloud/requirements.txt
streamlit
tensorflow-cpu==2.12.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
scikit-learn==1.2.2